main/pyproject.toml
[project]
name = "fabricctl"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pyyaml"]
